Courier update

+++ src/main/java/com/apiit/eirlss/sales_component/package_courier/CourierController.java
@@ -3,7 +3,11 @@ package com.apiit.eirlss.sales_component.package_courier;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.Collection;
 import java.util.Optional;
+
+import com.apiit.eirlss.sales_component.package_order.SalesOrder;
+import com.apiit.eirlss.sales_component.package_order.SalesOrderRepository;
 
 @RestController
 @RequestMapping(path = "courier/")
@@ -11,6 +15,9 @@ public class CourierController {
 
     @Autowired
     CourierRepository courierRepository;
+
+    @Autowired
+    SalesOrderRepository salesRepo;
 
     @GetMapping(path = "all")
     public Iterable<Courier> GetCouriers() {
@@ -41,11 +48,22 @@ public class CourierController {
     public Boolean DeleteCourier(@PathVariable String couid){
 
         Optional<Courier> c = courierRepository.findById(couid);
-        if(c.isPresent())
-            courierRepository.deleteById(couid);
+        if(c.isPresent()){
+            int count = 0;
+            Iterable<SalesOrder> sales = salesRepo.findAllByCourierId(c.get().getCourierId());
+            for(SalesOrder sale: sales){
+                count++;
+            }
 
-        System.out.println(c.get().getCourierId() + " is deleted");
+            if(count >= 1){
+                return false;
+            }else {
+                courierRepository.deleteById(couid);
+                return true;
+            }
+        }else {
+            return false;
+        }
 
-        return c.isPresent();
     }
 }
